Adopt Typescript. Add page objects.
This commit introduces two big changes.  First of all, we're using
Typescript instead of Babel now.  It's turned into something pretty
enjoyable, but mostly what we wanted was to measure the integration
pains.  There were some, but not tons.

Secondly, routing is now done via page objects.  In plain language, a
page object is a thing which can produce both a path (for links) and a
route object (for react router).  In other words, if you wanted to
create a new route you should do it with page objects.