Decouple components.
In preparation to add a Feathers back-end, we want to tease apart a lot
of the interlocking dependencies we had going on and shift to a more
dynamic model.

Not pictured:  Weeks, WEEKS of attempting to get this damned thing to
work universally.  The big problem with universal code is that I don't
know enough about both Typescript or Webpack to arrange a server side
build of it.

Anyways, onward.