Integrate Feathers, remove Typescript.
First of all, on removing Typescript:

Typescript is an important decision to make about a repo, and it's
probably not a great plan to just do it halfway.  Unfortunately, my own
knowledge (or, well, lack of knowledge) on its internals had turned into
a pretty endless time sink.

This all came to a head when implementing a Feathers API and integrating
it into the front end.  First of all, Feathers doesn't generate
Typescript code or run on it at all, and the process of converting
things over to Typescript was looking to be pretty expensive in terms of
time.

But, the good news is, Feathers is integrated, and there's a small
signup / login tool integration!

This is big, for a lot of reasons.  Foremost is that it creates a
websocket compliant concept of all potential data: or, in other words,
it means we can do a very realtime interface for all front-end code we
make in the repo.